Fixed two compiler warnings

+++ src/detection/displayserver/xcb.c
@@ -175,7 +175,7 @@ typedef struct XcbRandrData
 static bool xcbRandrHandleModeInfo(XcbRandrData* data, xcb_randr_mode_info_t* modeInfo)
 {
     uint32_t refreshRate = ffdsParseRefreshRate((int32_t) (
-        modeInfo->dot_clock / (modeInfo->htotal * modeInfo->vtotal)
+        modeInfo->dot_clock / (uint32_t) (modeInfo->htotal * modeInfo->vtotal)
     ));
 
     return ffdsAppendResolution(
